+++ autofit/database/aggregator/__init__.py
@@ -1,0 +1,1 @@
+from .aggregator import Aggregator
